isolate halo_patch_exchanger for ExchangeAndMerge
